feat(ResultPrinter): 구입한 로또 정보를 출력한다
ResultPrinter에 추가된 메서드:
- printLottoesBought 메서드를 통해 구입한 로또의 개수와, 로또들의 번호를 출력한다.
- 출력은 1줄씩 줄바꿈하여 출력된다.

+++ src/main/java/lotto/generator/LottoGenerator.java
@@ -16,7 +16,7 @@ public class LottoGenerator {
         return new Lotto(lottoNumbers);
     }
 
-    public List<Lotto> generate(int numberOfPurchases) {
+    public List<Lotto> generate(long numberOfPurchases) {
         List<Lotto> lottoes = new ArrayList<>();
         for (int i = 0; i < numberOfPurchases; i++) {
             lottoes.add(generate());
